refactor into deb_packages.cpp
